Allow storing more than one GUID for each metainfo file
To allow this, remove the constraint on the fwid,guid which depends on the
previous patch in this branch.

Fixes: https://github.com/hughsie/lvfs-website/issues/3

To migrate:
> ALTER TABLE firmware_md DROP INDEX id;
> ALTER TABLE firmware_md ADD CONSTRAINT id UNIQUE(fwid,metainfo_id);
> ALTER TABLE firmware_md MODIFY guid TEXT DEFAULT NULL;

If one day we need to query the database by GUID we'll have to also add
something like:

CREATE TABLE firmware_md_guid (
  metainfo_id VARCHAR(40) DEFAULT NULL,
  guid VARCHAR(36) DEFAULT NULL,
  UNIQUE KEY id (fwid,metainfo_id)
) CHARSET=utf8;

...and also remove the `guid` column from firmware_md (also doing a dbmigrate)
but at the moment we're always getting all the columns from all the firmware so
we can encode the data in the database however we like.