Fix rebuilding the embargoed metadata
The firmware filtering wasn't working correctly when the group_id was None.
To fix, just make things simpler so that we just give _metadata_update_group()
each group ID in turn.

As regenerating the metadata isn't a common operation, it doesn't really matter
that we block for a few seconds while processing or that the affidavit is
re-created each time.