Use the sqlalchemy ORM rather than hand-crafted SQL
MySQL-python isn't the best maintained module in the world, and using it
directly means we can't switch in a different engine like SQLite for testing.

This also removes quite a lot of boilerplate code and allows us to fix up a few
historial schema mistakes.

To migrate the database do:

    FLASK_APP=app/__init__.py flask-2 modifydb