Show which reports match a specific issue
This lets us 'test' the conditions without actually enabling the issue.

We also have to be a little bit careful when deciding if the QA user can view
the raw JSON -- so do a lookup on the firmware group_id and show missing text
if that fails the usual tests.