Allow vendors to hide the direct download link in the search results
For arguable reasons, some vendors want to hide the direct download links for
firmware on the LVFS pages. This means the firmware files are only available to
download by programs that parse the XML metadata, for instance gnome-software
and fwupdmgr.

To migrate, do:

    ALTER TABLE firmware ADD inhibit_download TINYINT DEFAULT 0;

To use this in firmware files, include a section in the metainfo.xml as follows:

    <custom>
      <value key=LVFS::InhibitDownload/>
    </custom>