Fix deleting firmware
Now we have a foreign key constraint on the Client object we also need to make
sure that none of the downloads are left 'dangling'.

Of course, deleting firmware is not something to recommend as it also deletes
any firmware events, the client history and all the statistics have to change.