Use ACLs for different actions that require authentication
We can put all the permission logic in models.py in the correct classes to
reduce the amount of duplicated logic in the templates and the actual method
permissions checks.

By suppliying /what/ we're trying to do we can make changes in one place without
having to understand where else in the code the permission tests are being done.