Introduce another user attribute for testing and stable approval
QA used to mean two things:

 * Able to push to testing and stable
 * Can see all firmware uploaded to the group

This means that any vendor team trialling the LVFS cannot see each others
firmware while waiting for legal assurance that the firmware is re-distributable
on the LVFS. It doesn't make sense to tie these things together, and so add a
new user attribute for the approval to push to testing and stable.

To migrate, set the approved attribute on existing users with QA access.