Stack the testing and stable metadata in the various vendor graphs
It was pointed out that the numbers didn't match what was found in the stable
metadata, which is somewhat unsatisfactory for public stats like this.

Show the firmware in the testing state as a stacked bar on top of the stable
firmware, and sort the vendors by the stable+testing firmware count.