Include the component ID in the URL query string in all cases
At the moment users can add a requirement or keyword to a firmware, but in the
event log we don't know which component they've modified as it only stores the
URL, not all the POST data.

Make the URL more RESTful by including the component ID to easily fix this.