Store the component description and update text as Markdown
It's become very obvious that allowing non-technical users to modify the update
XML directly was a bad idea. Quite a few of the update descriptions are not
valid XML (e.g. '<p><p></p>') and ODMs seem to just be pasting in the
somewhat-markdown-ish release notes into the component page and expecting
AppstreamGlib to do the right thing, which it doesn't do.

Convert the existing release notes to Markdown format, and give an example in
the component update page about what makes a good update description.