Do not traceback when the component install duration is invalid
The root cause of this failure is that firmware components in the database that
were adjusted to include an install duration had the 'missing' value set as
NULL, rather than zero. This caused any modification to the update description
to fail.

Do two things, the first being fixing up the database to convert NULL to 0, and
the second to make the conversion to integer more resilient.