Correctly fallback to the vendor version format if unspecified
The existing code was correctly falling back to the desired version format for
display in templates, but was not falling back when generating metadata.
Make things simpler by just setting the vendor version format at upload time
rather than falling back in all the places we need the component version format.

Also, migrate the existing firmware that is not a semver, has a vendor version
format, and no component version format. Ideally vendors would include the
version format in the metadata, but if we supply the fallack feature on the
LVFS, it probably ought to work...

Fixes https://github.com/hughsie/fwupd/issues/854